fix: :truck: Moved tools function to lib/tools.ts
Issue: #138
